% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{corner}
\alias{corner}
\title{Utility Functions}
\usage{
corner(table, nrow, ncol)
}
\description{
Perform misc utility functions
* `corner(table, nrow=5, ncol=5)` show the the top left corner of a data.frame/tibble/matrix.
* `min_nonzero()` Find the smallest non-zero, non-NA value in a vector.
}
