% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_golay.R
\name{decode_golay}
\alias{decode_golay}
\title{Functions for decoding golay barcodes}
\usage{
decode_golay(barcodes)
}
\arguments{
\item{barcodes}{a vector of barcodes to be corrected}
}
\value{
a data.table with the following columns: original_bc, corrected_bc, n_errors
}
\description{
This is a port of the original QIIME2 python code from q2_demux/_ecc.py into R made by Cecilia Knoecker 2020/11/17 and then added into qiime2R by JB
}
