% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features.R
\name{filter_features}
\alias{filter_features}
\title{Filter low abundance features from a table.}
\usage{
filter_features(features, minsamples, minreads, verbose)
}
\arguments{
\item{features}{Table of feature/OTU/ASV counts where Samples are columns and feature IDs are row names}

\item{minsamples}{A minimum number of samples (>=) for the feature/OTU/ASV to be observed in. (Default=2)}

\item{minreads}{A minimum number of reads  (>=) for a feature/OTU/ASV to be kept across all samples. (Default=2)}

\item{verbose}{Should summary be printed? (T/F)}
}
\value{
filtered feature table
}
\description{
Takes a table of features (samples are columns) and remove features that do not appear at in at least X samples OR do not have at least Y reads.
}
