% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qza.R
\name{import_qza}
\alias{import_qza}
\alias{print.qiime2_qza}
\alias{print.qza_metadata}
\title{Read QIIME2 Artifacts (.qza)}
\usage{
import_qza(file, tmp = tempdir(), rm = TRUE)

\method{print}{qiime2_qza}(x, ...)

\method{print}{qza_metadata}(x, ...)
}
\arguments{
\item{file}{(string) Path to the input file,
e.g.: \code{file = "~/data/moving_pictures/table.qza"}.}

\item{tmp}{(string) A temporary directory that the object will be
decompressed to (\code{default = "tempdir()"}).}

\item{rm}{(\code{TRUE}|\code{FALSE})
Should the decompressed object be removed at completion of function
(default = \code{TRUE}).}
}
\value{
An R object with data and metadata in QIIME artifact file (.qza) as
\code{list}, \code{data.frame}, or \code{matrix} (depending on the type of artifact).
The object has given 2 additional classes:
\itemize{
\item Class \code{qiime2_qza};
\item Other class which coincides with QIIME2 semantic type, e.g.
\code{FeatureData[Sequence]}.
}

The object contains attribute \code{metadata}, which is a named list of the
following fields:
\itemize{
\item \code{data} -- the raw data, e.g., OTU table as matrix or tree in phylo format;
\item \code{uuid} -- the unique identifer of the artifact;
\item \code{type} -- the semantic type of the object (e.g., \code{FeatureData[Sequence]});
\item \code{format} -- the format of the qiime artifact;
\item \code{provenance} -- information tracking how the object was created;
\item \code{contents} -- a table of all the files contained within the artifact and
their file size;
\item \code{version} -- the reported version for the artifact, a warning error may be
thrown if a new version is seen;
}
}
\description{
Reads QIIME2 artifacts (.qza) file and extracts embedded data and object
metadata into an R object.
}
\examples{
\dontrun{
asv <- import_qza("data/table.qza")

# Get metadata
attr(asv, "metadata")
}

}
