% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_table.R
\name{interactive_table}
\alias{interactive_table}
\title{Display an interactive table}
\usage{
interactive_table(table, nrow)
}
\arguments{
\item{table}{a data.frame/tibble/matrix}

\item{nrow}{number of rows to display in table (default=10)}
}
\description{
Uses DT:datatable to embed an interactive table with export options in an R session or markdown document
}
\examples{
\dontrun{interactive_table(metadata)}
}
