% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_q2metadata.R
\name{is_q2metadata}
\alias{is_q2metadata}
\title{checks if metadata is in qiime2 (.tsv)}
\usage{
is_q2metadata(file)
}
\arguments{
\item{file}{path to the input file, ex: file="~/data/moving_pictures/table.qza"}
}
\value{
TRUE/FALSE
}
\description{
Checks to see if a file is in qiime2 metadata format, ie contains #q2:types line dictating the type of variable (categorical/numeric)
}
\examples{
\dontrun{metadata<-is_q2metadata("q2metadata.tsv")}
}
