% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizers.R
\name{make_clr}
\alias{make_clr}
\title{Normalization Functions}
\usage{
make_clr(features, prior, czm)
}
\arguments{
\item{features}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names. Expects matrix or data.frame.}

\item{prior}{Only relevant to CLR: A numeric value to add before log transformation (default 0.5). *Ignored if CZM=TRUE}

\item{czm}{Only relevant to CLR: Should count zero multiplicative method be used instead of adding simple prior (TRUE/FALSE, defaults to FALSE)}
}
\value{
Table of normalized abundances
}
\description{
Takes a numeric table and transforms to a number of options as below. Note: Rows are expected to be features and columns are expected to be sample IDs.
* `make_clr()` A centered log2-ratio (columns sum to ~0). See and cite 10.1139/cjm-2015-0821
* `make_percent()` Convert to percent (columns sum to 100)
* `make_proportion()` Convert to proportion (columns sum to 1)
}
