% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ordination.R
\name{parse_ordination}
\alias{parse_ordination}
\title{Parse an imported q2 ordination text file}
\usage{
parse_ordination(artifact, tmp)
}
\arguments{
\item{artifact}{takes an artifact containing the unparsed ordination format (result of readLines)}

\item{tmp}{the tmp dir for artifact}
}
\value{
a data.frame with feature IDs as row names and the columns: Kingdom, Phylum, Class, Order, Family, Genus, Species
}
\description{
Parse an imported q2 ordination text file
}
\examples{
\dontrun{ord<-parse_ordination(ordination_artifact, tmpdir)}
}
