% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_taxonomy.R
\name{parse_taxonomy}
\alias{parse_taxonomy}
\title{Parse Q2 taxonomy}
\usage{
parse_taxonomy(taxonomy, tax_sep, trim_extra)
}
\arguments{
\item{taxonomy}{the imported taxonomy object (the result of read_qza() containing columns Feature.ID, Taxon, and Confidence)}

\item{tax_sep}{The separator between taxonomic levels. Defaults to one compatible with both GreenGenes and SILVA ("; " OR ";")}

\item{trim_extra}{Remove leading characters from taxonomic levels: ex: k__ or D_0__. TRUE/FALSE. default=TRUE}
}
\value{
a data.frame with feature IDs as row names and the columns: Kingdom, Phylum, Class, Order, Family, Genus, Species
}
\description{
Parse Q2 taxonomy
}
\examples{
\dontrun{taxonomy<-parse_taxonomy(taxonomy_artifact)}
}
