% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_provenance.R
\name{plot_provenance}
\alias{plot_provenance}
\title{plot the provenance of a QIIME2 artifact (.qza)}
\usage{
plot_provenance(artifact)
}
\arguments{
\item{artifact}{the object returned by read_qza}
}
\value{
a ggplot object which can be printed to the screen and/or modified
}
\description{
extracts embedded data and object metadata into an R session
}
