% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_provenance.R
\name{print_provenance}
\alias{print_provenance}
\title{plot the provenance of a QIIME2 artifact (.qza)}
\usage{
print_provenance(artifact)
}
\arguments{
\item{artifact}{the object returned by read_qza}
}
\value{
a text summary for the provenance information
}
\description{
extracts embedded data and object metadata into an R session
}
\examples{
\dontrun{print_provenance(artifact)}
}
