% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiime2R-package.R
\docType{package}
\name{qiime2R-package}
\alias{qiime2R}
\alias{qiime2R-package}
\title{qiime2R: qiime2R}
\description{
Importing QIIME2 artifacts and associated data into R sessions.
}
\keyword{internal}
