% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qza_to_phyloseq.R
\name{qza_to_phyloseq}
\alias{qza_to_phyloseq}
\title{generates a phyloseq (.qza)}
\usage{
physeq<-qza_to_phyloseq(features="table.qza", tree="rooted-tree.qza", taxonomy="taxonomy.qza", metadata="sample-metadata.tsv")
}
\arguments{
\item{features}{file path for artifact containing a feature (OTU/SV) table}

\item{tree}{file path for  artifact containing a tree}

\item{taxonomy}{file path for artifact containg taxonomy}

\item{metadata}{file path for a qiime2-compliant TSV metadata file}
}
\value{
a phyloseq object
}
\description{
Construct a phyloseq object from multiple qiime2 artifacts (.qza). Embedded metadata for provenance is not maintained in this function and instead read_qza() should be used.
}
