% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qza_to_tse.R
\name{qza_to_tse}
\alias{qza_to_tse}
\title{Generates a TreeSummarizedExperiment (TSE) object from .qza artifacts
\code{\link[TreeSummarizedExperiment:SummarizedExperiment-class]{TreeSummarizedExperiment}}}
\usage{
qza_to_tse(features, tree, taxonomy, metadata, tmp)
}
\arguments{
\item{features}{file path for artifact containing a feature (OTU/SV) table}

\item{tree}{file path for artifact containing a tree}

\item{taxonomy}{file path for artifact containg taxonomy}

\item{metadata}{file path for a qiime2-compliant TSV metadata file}

\item{tmp}{a temporary directory that the object will be decompressed to.}
}
\value{
a TreeSummarizedExperiment object
\code{\link[TreeSummarizedExperiment:SummarizedExperiment-class]{TreeSummarizedExperiment}}
}
\description{
Construct a TreeSummarizedExperiment object from multiple 
qiime2 artifacts (.qza). 
Embedded metadata for provenance is not maintained in this function and 
instead \code{read_qza()} should be used.
}
\examples{
# (Data is from tutorial
https://docs.qiime2.org/2020.2/tutorials/moving-pictures/)

tse <-qza_to_tse(
features="path_to_table.qza",
tree="path_to_rooted-tree.qza",
taxonomy="path_to_taxonomy.qza",
metadata = "path_to_sample-metadata.tsv"
)

}
