% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_q2biom.R
\name{read_q2biom}
\alias{read_q2biom}
\title{read qiime2 biom file (version 2.1)}
\usage{
read_q2biom(file)
}
\arguments{
\item{file}{path to the input file, ex: file="~/Downloads/3372d9e0-3f1c-43d8-838b-35c7ad6dac89/data/feature-table.biom"}
}
\value{
a matrix of values
}
\description{
Loads a version 2.1 spec biom file (http://biom-format.org/documentation/format_versions/biom-2.1.html) as expected to be found within a qiime2 artifact.
}
\examples{
\dontrun{metadata<-read_q2biom("feature-table.biom")}
}
