% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qza.R
\name{read_qza}
\alias{read_qza}
\title{read qiime2 artifacts (.qza)}
\usage{
read_qza(file, tmp, rm)
}
\arguments{
\item{file}{path to the input file, ex: file="~/data/moving_pictures/table.qza"}

\item{tmp}{a temporary directory that the object will be decompressed to (default="/tmp")}

\item{rm}{should the decompressed object be removed at completion of function (T/F default=T)}
}
\value{
a named list of the following objects:
\itemize{
\item artifact$data - the raw data ex OTU table as matrix or tree in phylo format
\item artifact$uuid - the unique identifer of the artifact
\item artifact$type - the semantic type of the object (ex FeatureData[Sequence])
\item artifact$format - the format of the qiime artifact
\item artifact$provenance - information tracking how the object was created
}
}
\description{
extracts embedded data and object metadata into an R session
}
