% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_table.R
\name{subsample_table}
\alias{subsample_table}
\title{subsample_table}
\usage{
subsample_table(features, depth, seed, with_replace, verbose)
}
\arguments{
\item{features}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{depth}{Count depth, defaults to min(colSums(OTUTABLE)). Note, samples with less than the desired number of read will be removed!}

\item{seed}{A randomization SEED, defaults to 182. This is ignored for multiple subsamples and instead the seeds 1:NSAMPS is used.}

\item{with_replace}{Should subsampling be with or without replacement? (TRUE/FALSE) Default=FALSE to be analogous to qiime2; however, this may take longer to calculate. Please see phyloseq::rarefy_even_depth for more information.}

\item{verbose}{Should progress and metrics be printed to screen via message()? Default=TRUE}
}
\value{
Subsampled Table
}
\description{
This function has been ported from MicrobeR. Takes a Feature/OTU/ASV table of counts where samples are columns and feature names are row names and returns a table with even coverage on a per-sample basis. Now this function is really just an alias for rarefy_even_depth from phyloseq to simplify its usage for non-phyloseq objects. NOTE: Sampling with replacement for single rarefraction method!
}
