% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_taxa.R
\name{summarize_taxa}
\alias{summarize_taxa}
\title{summarize_taxa}
\usage{
summarize_taxa(features, taxonomy)
}
\arguments{
\item{features}{a table of abundances where rows are features and columns are samples}

\item{taxonomy}{a table of taxonomies that has been parsed (ex using parse_taxonomy). Should contain the following columns: Kingdom,Phylum,Class,Order,Family,Genus,Species.}
}
\value{
a named list of abundances aggregated at each level.
}
\description{
Takes a feature table (ex OTUs/SVs) and a parsed taxonomy table to generate a named list of feature tables with abundances summed to each taxonomic level.
}
\examples{
\dontrun{taxasums<-summarize_taxa(svtable, taxonomy); taxasums$Species}
}
