% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_barplot.R
\name{taxa_barplot}
\alias{taxa_barplot}
\title{taxa_barplot}
\usage{
taxa_barplot(features, metadata, "treatment")
}
\arguments{
\item{features}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{metadata}{A Table of metadata where sample names are the row names OR there is already a column with the name "SampleID"}

\item{category}{A metadata category to block samples by which is a column in the metadata (faceting via ggplot2)}

\item{normalize}{How should table be normalized? (default="percent", options=c("percent","proportion","none"))}

\item{ntoplot}{A number of features to plot.}
}
\value{
Barplot
}
\description{
Uses ggplot2 to create a stacked barplot, for example on phylum level abundances. The most abundant features (defaults to 10, based on rowMeans) will be plotted unless user specified.
}
