% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_heatmap.R
\name{taxa_heatmap}
\alias{taxa_heatmap}
\title{taxa_heatmap}
\usage{
taxa_heatmap(features, metadata, "treatment")
}
\arguments{
\item{features}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{metadata}{A Table of metadata where sample names are the row names OR there is already a column with the name "SampleID"}

\item{category}{A metadata category to block samples by which is a column in the metadata (faceting via ggplot2)}

\item{normalize}{How should table be normalized?  default="log10(percent)", options=c("log10(percent)","clr","none"). Note if using  log10(percent), a pseudocount 1 read is added before conversion to percent}

\item{ntoplot}{A number of features to plot.}
}
\value{
Barplot
}
\description{
Uses ggplot2 to create a heatmap, for example on phylum level abundances. The most abundant features (defaults to 10, based on rowMeans) will be plotted unless user specified.
}
