% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_q2r.R
\name{theme_q2r}
\alias{theme_q2r}
\title{A ggplot2 theme for publication-style figures}
\usage{
theme_q2r()
}
\description{
A ggplot2 theme for publication-style figures
}
\examples{
\dontrun{ggplot(data) + theme_q2r()}
}
