% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_q2manifest.R
\name{write_q2manifest}
\alias{write_q2manifest}
\title{Generates a  read manifest for importing sequencing data into qiime2}
\usage{
write_q2manifest(outfile, directory, extension, paired, Fwd, Rev)
}
\arguments{
\item{outfile}{filename for output (default: manifest_[timestamp].txt)}

\item{directory}{directory containing reads}

\item{extension}{file extension (default: .fastq.gz)}

\item{paired}{are reads in paired format? TRUE/FALSE (default=FALSE)}

\item{Fwd}{string used to denote a forward read (default= _R1)}

\item{Rev}{string used to denote a reverse read (default= _R2)}
}
\description{
Scans a directory for files with matching sequencing data (default: fastq.gz) and then generates a q2 compliant manifest.
}
\examples{
\dontrun{write_q2manifest("q2manifest.txt","/yourdirhere/reads/", extension=".fastq.gz", paired=TRUE)}
}
