% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{corner}
\alias{corner}
\title{Utility Functions}
\usage{
corner(table, nrow, ncol)
}
\arguments{
\item{corner()}{show the the top left corner of a data.frame/tibble/matrix.}

\item{min_nonzero()}{Find the smallest non-zero, non-NA value in a vector.}
}
\description{
Perform misc utility functions:
}
